This commit:  changes the Noise class to allow independent frequency bins for each stored PSD, extracts the simulation of one timestream into a function that can be better unit-tested and incorporated into multiple noise simulation functions, improves the frequency spacing of the AnalyticNoise class, and extends the unit tests to have a better variety of knee frequencies.

+++ tests/test_psd_math.py
@@ -31,7 +31,7 @@ class PSDTest(MPITestCase):
         self.sigma = 10.
         self.signal = np.random.randn(self.nsamp) * self.sigma
         self.flags = np.zeros(self.nsamp, dtype=np.bool)
-        self.flags[self.nsamp/4:self.nsamp/2] = True
+        self.flags[int(self.nsamp/4):int(self.nsamp/2)] = True
 
 
     def test_autocov_psd(self):
